+++ app/src/main/java/com/swent/skillswap/model/post/PostFirestoreRepository.kt
@@ -25,7 +25,7 @@ class PostFirestoreRepository(db: FirebaseFirestore) : PostRepository {
     }
 
     override suspend fun getMultiplePosts(
-        numberOfPosts: Long,
+        numberOfPosts: Int,
         type: PostType,
         titleContains: String,
         ownerId: String,
